+++ docs/changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v1.0.0
+## v0.2.0
 
 * Rename `retryhttp.retry_http_errors` to [`retryhttp.retry.retry`][].
 * Rename `retryhttp.wait_http_errors` to [`retryhttp.wait_context_aware`][].